chore: add bare-minimum logging support
Follow the Python documentation guidelines for "Configuring Logging
for a Library" [1]

Which is basically adding these two lines:
  import logging
  logging.getLogger(__name__).addHandler(logging.NullHandler())

Setup a very basic usage of logging in `gitlab/client.py`

By using the NullHandler it means that by default any log messages
output will not be displayed. It is up to the client application to do
a `logging.basicConfig()` call to get log messages to display.

[1] https://docs.python.org/3/howto/logging.html#configuring-logging-for-a-library
Related: #2080